Extract shared OAuth helper, move ExtraConfigType to types
- Extract postOAuthMessageAndClose helper to deduplicate
  postMessage/BroadcastChannel/close logic between setup.tsx
  and finalize.tsx
- Move ExtraConfigType and ExtraConfigTypeSchema from page file
  to types/oauth/lib.ts
- Update all provider imports to use @app/types
