+++ front/lib/api/oauth.ts
@@ -36,8 +36,8 @@ import { ZendeskOAuthProvider } from "@app/lib/api/oauth/providers/zendesk";
 import { finalizeUriForProvider } from "@app/lib/api/oauth/utils";
 import type { Authenticator } from "@app/lib/auth";
 import logger from "@app/logger/logger";
-import type { ExtraConfigType } from "@app/pages/w/[wId]/oauth/[provider]/setup";
 import type {
+  ExtraConfigType,
   OAuthAPIError,
   OAuthConnectionType,
   OAuthProvider,
