+++ front/pages/oauth/[provider]/finalize.tsx
@@ -2,8 +2,8 @@ import type { InferGetServerSidePropsType } from "next";
 import { useEffect } from "react";
 
 import { makeGetServerSidePropsRequirementsWrapper } from "@app/lib/iam/session";
+import { postOAuthMessageAndClose } from "@app/lib/oauth/postOAuthMessageAndClose";
 import { useFinalize } from "@app/lib/swr/oauth";
-import logger from "@app/logger/logger";
 import type { OAuthProvider } from "@app/types";
 import { isOAuthProvider } from "@app/types";
 
@@ -42,53 +42,19 @@ export default function Finalize({
       // (e.g. go to GitHub and configure repositories from Dust App in Settings, hence no opener).
       const res = await doFinalize(provider, queryParams);
 
-      // Prepare message data
-      const messageData = res.isErr()
-        ? {
-            type: "connection_finalized",
-            error: res.error.message || "Failed to finalize connection",
-            provider,
-          }
-        : {
-            type: "connection_finalized",
-            connection: res.value,
-            provider,
-          };
-
-      // Method 1: window.opener (preferred, direct communication)
-      if (window.opener && !window.opener.closed) {
-        try {
-          window.opener.postMessage(messageData, window.location.origin);
-        } catch (e) {
-          logger.error(
-            { err: e },
-            "[OAuth Finalize] window.opener.postMessage failed",
-            e
-          );
-        }
-      } else {
-        // Method 2: BroadcastChannel (fallback for modern browsers)
-        try {
-          const channel = new BroadcastChannel("oauth_finalize");
-          channel.postMessage(messageData);
-          setTimeout(() => channel.close(), 100);
-        } catch (e) {
-          logger.error(
-            { err: e },
-            "[OAuth Finalize] BroadcastChannel failed",
-            e
-          );
-        }
-      }
-
-      // Close window after a short delay to ensure message delivery
-      setTimeout(() => {
-        window.close();
-        // If window.close() fails, redirect to home
-        setTimeout(() => {
-          window.location.href = window.location.origin;
-        }, 100);
-      }, 1000);
+      postOAuthMessageAndClose(
+        res.isErr()
+          ? {
+              type: "connection_finalized",
+              error: res.error.message || "Failed to finalize connection",
+              provider,
+            }
+          : {
+              type: "connection_finalized",
+              connection: res.value,
+              provider,
+            }
+      );
     }
     void finalizeOAuth();
   }, [queryParams, provider, doFinalize]);
