+++ front/types/oauth/lib.ts
@@ -3,6 +3,9 @@ import * as t from "io-ts";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { validateUrl } from "@app/types/shared/utils/url_utils";
 
+export const ExtraConfigTypeSchema = t.record(t.string, t.string);
+export type ExtraConfigType = t.TypeOf<typeof ExtraConfigTypeSchema>;
+
 export const OAUTH_USE_CASES = [
   "connection",
   "labs_transcripts",
